cig-kg-backend-master/requirements.txt
langchain
python-dotenv
neo4j
tqdm
pydantic
fastapi
pathlib
pymupdf4llm
pypdf2
openai
uvicorn
langchain_community
python-multipart